update id_ to int

+++ caches-intelligent/src/main/java/com/dflc/cache/iot/entity/Device.java
@@ -9,7 +9,7 @@ public class Device extends Cachedable {
 
     protected Long id;
 
-    protected Long id_;
+    protected Integer id_;
 
     protected String name;
